requirements.txt

django==4.2.4
pillow==10.0.0
phonenumbers==8.13.33
django-phonenumber-field==7.3.0